Implemented streaming of static assets
It is obviously inefficient to read large files entirely into memory in
order to send them to a downloading client. Therefore, the mechanism has
been changed to that of an async stream.

Streams are not really what we want - we want progressive reading and
writing of the file - but there is not an obvious way with Axum/Tokio to
achieve this. All the documentation talks about streaming. One major
usability downside to this is that we no longer send the file size - and
there is also a performance impact.

Testing shows that this approach, while using a minimum amount of
memory, is many times slower than the previous method - significantly
so. This can be partially resolved with some later amends.

  - Changed get_static_asset() to return a StreamBody.