Added Dockerfile
Added a Dockerfile, backported from Rustmark, to help with building the
application in a container. This builds the application in two stages -
first it creates a build environment in which to prepare everything, and
then the final binary is transferred to a clean, minimal environment
which is suitable for distribution.

When evaluating options when testing on Rustmark, debian:bullseye-slim
resulted in a 94MB image size (85MB once the executable was run through
UPX), distroless gave a 36MB image size (27MB once run through UPX), and
Alpine give a 21MB image (12MB after running the executable through
UPX). There do not currently appear to be any side-effects from using
Alpine and musl, but it should be noted that as it does not use glibc,
some crates may not be compatible.

Mold is used as the linker, and UPX is used to minimise the size of the
final executable. This reduces the binary from 13MB down to 4MB. UPX is
implemented as its own layer, as it is worth having an extra layer to
gain the advantage of easily toggling it on and off without rebuilding
the whole application.

The Config.toml file is added right at the end, directly to the runnable
image instead of to the build image. This means that a change of
configuration will not trigger a rebuild.

Some aspects of optimisation are useful to control by passing arguments
to docker build. The default position is to optimise for the smallest
possible image size, but this can result in longer compilation times.
Although this Dockerfile is intended for production builds, in some
cases speed may be favoured. Most of the options affecting this are
configured through the Cargo.toml file, but notably the use of UPX does
have a significant effect on time, and therefore an option has been
added to specifically control it. Additionally, it can be useful to
adjust build settings through command arguments, and therefore another
argument has been added to pass options to cargo build.

It is useful to be able to build for dev and other profiles as well as
release, so an argument has been added for this. UPX is not used for the
dev profile.

Based on the results of testing on Rustmark, some slight changes have
been made to the default release build profile. These changes are
notably tuned for the static musl Docker environment, and have different
effects if building in a GNU or non-musl environment. The focus is
establishing a sensible default position to balance obtaining the
smallest possible deployable image size versus the fastest possible
initial build and iterative rebuild speed.

Cargo options   Dev   Release
codegen-units:  256   16       Trade-off of speed vs optimisation
debug:          2     0        Debug symbols: 0/1/2
lto:            off   thin     Link-time optimisation: off/thin/fat
opt-level:      0     3        Optimise for size or speed: 0/1/2/3/s/z
strip:          none  symbols  Strip from binary: none/debuginfo/symbols

The summary of the changes is:

  - opt-level changed from z to 3 - back to the default value.
  - lto changed from true (fat) to thin. The default is false (no LTO).

There is no change to the defaults for building dev (debug) versions.

For more information on the tuning of these settings, refer to the
commits on the Rustmark project.

It's worth noting that the host IP to serve on needs to be set to
0.0.0.0 to allow outside traffic to connect. In other words, the host
entry in Config.toml should be set to "0.0.0.0" for a Docker setup.

  - Added Dockerfile backported from Rustmark.

  - Tweaked the release build profile based on Rustmark testing.

  - Updated README file with details of building and running a Docker
    container using the provided Dockerfile.